Added unit tests to sales module

+++ src/Modules/Sales/ECommerce.Modules.Sales.Application/Orders/Commands/Handlers/DeleteOrderItemHandler.cs
@@ -1,11 +1,6 @@
 ﻿using ECommerce.Modules.Sales.Application.Orders.Exceptions;
 using ECommerce.Modules.Sales.Domain.Orders.Repositories;
 using ECommerce.Shared.Abstractions.Commands;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ECommerce.Modules.Sales.Application.Orders.Commands.Handlers
 {
@@ -20,11 +15,16 @@ namespace ECommerce.Modules.Sales.Application.Orders.Commands.Handlers
 
         public async Task HandleAsync(DeleteOrderItem command)
         {
-            var orderItem = await _orderItemRepository.GetAsync(command.OrderItemId);
+            var orderItem = await _orderItemRepository.GetDetailsAsync(command.OrderItemId);
 
             if (orderItem is null)
             {
                 throw new OrderItemNotFoundException(command.OrderItemId);
+            }
+
+            if (orderItem.Order is not null)
+            {
+                throw new OrderItemCannotBeDeletedException(command.OrderItemId);
             }
 
             await _orderItemRepository.DeleteAsync(orderItem);
